chore(weave): move eval worker out to weave.eval_worker (decouple server)
The trace server reached into the client SDK only through the eval-execution
path: evaluate_model_worker.py and isolated_client_executor.py import the client
(Evaluation, LLMAsAJudgeScorer, WeaveClient, refs, weave_client_context) to run
user evaluations in a subprocess. They lived under weave/trace_server/, so
importing them — even just the EvaluateModelDispatcher seam the server needs —
transitively pulled the client into the server.

Move both modules to a new sibling package weave/eval_worker/. It may import
BOTH sides (it is never a contract source) — it is the client-coupled glue that
runs evals out-of-process — while a new import-linter contract ("Eval worker
must not be imported by the trace server or client SDK") forbids anything in the
trace server or client SDK from importing it, so the decoupling cannot silently
regress through the new package. The trace server now depends only on a
client-free seam, weave.trace_server.eval_model_dispatcher.EvaluateModelDispatcher
(an ABC over EvaluateModelArgs), which the ClickHouse and in-memory servers
import for their constructor param. The concrete dispatcher (core repo) is
injected at runtime as before.

No back-compat shims: the old weave.trace_server paths are deleted outright, and
the now-empty weave/trace_server/workers/ package is removed along with the
redundant "Workers must not import the Weave client SDK" contract (the
trace-server contract already covers all of weave.trace_server). The core repo's
three consumers are repointed to weave.eval_worker.* in lockstep (companion PR
wandb/core#46453); because the old paths are gone, the two changes land together
and the core weave-submodule pin must include this commit.

In-repo tests are repointed to weave.eval_worker.*. Removes all 7
trace_server->client eval-worker/isolated ignores. The only remaining
trace-server->client ignores are the 4 llm_structured_model rows (deferred: that
file is a runnable weave.Model destined to move client-side via codegen).
