+++ tests/trace_server/isolated_client_executor/test_isolated_client_executor.py
@@ -20,14 +20,14 @@ import weave
 from tests.trace_server.isolated_client_executor.cross_process_trace_server import (
     CrossProcessTraceServerReceiver,
 )
-from weave.trace.context.weave_client_context import get_weave_client
-from weave.trace.ref_util import get_ref
-from weave.trace.weave_client import WeaveClient
-from weave.trace_server.isolated_client_executor import (
+from weave.eval_worker.isolated_client_executor import (
     IsolatedClientExecutor,
     IsolatedClientExecutorError,
     IsolatedClientExecutorTimeoutError,
 )
+from weave.trace.context.weave_client_context import get_weave_client
+from weave.trace.ref_util import get_ref
+from weave.trace.weave_client import WeaveClient
 from weave.trace_server.trace_server_interface import ObjQueryReq, TraceServerInterface
