+++ weave/trace_server/in_memory_trace_server.py
@@ -88,6 +88,9 @@ from weave.trace_server.errors import (
     RequestTooLarge,
     handle_clickhouse_query_error,
 )
+from weave.trace_server.eval_model_dispatcher import (
+    EvaluateModelDispatcher,
+)
 from weave.trace_server.feedback import (
     TABLE_FEEDBACK,
     format_feedback_to_res,
@@ -150,9 +153,6 @@ from weave.trace_server.ttl_settings import (
     invalidate_ttl_cache,
 )
 from weave.trace_server.validation import object_id_validator
-from weave.trace_server.workers.evaluate_model_worker.evaluate_model_worker import (
-    EvaluateModelDispatcher,
-)
 
 logger = logging.getLogger(__name__)
